// src/site/dto/site.request.dto.ts
/**
 * @Author: Gibeom Choi
 * @Date:   2023-06-02 17:31:44
 * @Last Modified by:   Gibeom Choi
 * @Last Modified time: 2023-06-02 17:33:55
 */
export class SiteRequestDTO {
  name: string;
  location: string;
  type_code: number;
  emergency_contact: string;
  start_date: Date;
  end_date: Date;
}